requirements/base.txt
bleach>=2.0.0
dj-database-url>=0.4.2
dj-static>=0.0.6
django-autoslug>=1.9
django-taggit>=0.22
Django==1.10
Markdown==2.6
Pillow==4.2
psycopg2==2.7
python-decouple==3
pytz==2017.2
Unipath==1.1
whitenoise==3.3.0
